create PaymentStepWrapper
