+++ app/webpacker/components/RegistrationsV2/Register/PaymentOverview.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useQuery } from '@tanstack/react-query';
-import { Message } from 'semantic-ui-react';
+import { Icon, Message } from 'semantic-ui-react';
 import getRegistrationPayments from '../api/payment/get/getRegistrationPayments';
 import Loading from '../../Requests/Loading';
 import I18n from '../../../lib/i18n';
@@ -23,8 +23,11 @@ export default function PaymentOverview() {
   }
 
   return (
-    <Message success>
-      {I18n.t('registrations.entry_fees_fully_paid', { paid: registration?.payment.payment_amount_human_readable })}
+    <Message success icon>
+      <Icon name="checkmark" />
+      <Message.Content>
+        {I18n.t('registrations.entry_fees_fully_paid', { paid: registration?.payment.payment_amount_human_readable })}
+      </Message.Content>
     </Message>
   );
 }
